app/requirements.txt
Django==4.1.1
gunicorn==20.1.0
psycopg2==2.9.3
django-filter==22.1
djangorestframework-simplejwt==5.2.0
Pygments==2.13.0
celery==5.2.7
aiohttp==3.8.1